# Speed up sharing-aware extractor and fix spelling