--- a/plot.py
+++ b/plot.py
@@ -31,8 +31,8 @@ def process(js, extractors=[]):
     assert len(extractors) == 2
     e1, e2 = extractors
 
-    e1_cummulative=0
-    e2_cummulative=0
+    e1_cumulative=0
+    e2_cumulative=0
 
     summaries = {}
 
@@ -45,8 +45,8 @@ def process(js, extractors=[]):
             dag_ratio = d[e1]["dag"] / d[e2]["dag"]
             micros_ratio = max(1, d[e1]["micros"]) / max(1, d[e2]["micros"])
             
-            e1_cummulative += d[e1]["micros"];
-            e2_cummulative += d[e2]["micros"];
+            e1_cumulative += d[e1]["micros"];
+            e2_cumulative += d[e2]["micros"];
             
             summaries[name] = {
                 "tree": tree_ratio,
@@ -57,8 +57,8 @@ def process(js, extractors=[]):
             print(f"Error processing {name}")
             raise e
 
-    print(f"Cummulative time for {e1}: {e1_cummulative/1000:.0f}ms")
-    print(f"Cummulative time for {e2}: {e2_cummulative/1000:.0f}ms")
+    print(f"cumulative time for {e1}: {e1_cumulative/1000:.0f}ms")
+    print(f"cumulative time for {e2}: {e2_cumulative/1000:.0f}ms")
 
     print(f"{e1} / {e2}")